refactor: mock data file moved doc: added Swagger yaml
