+++ product-service/handlers/db/getProductByIdFromDB.ts
@@ -1,4 +1,4 @@
-import ProductsListMock from '../productsList.mock.json';
+import ProductsListMock from './productsList.mock.json';
 import { Product } from '../models/Product';
