+++ product-service/handlers/db/getProductsListFromDB.ts
@@ -1,4 +1,4 @@
-import ProductsListMock from '../productsList.mock.json';
+import ProductsListMock from './productsList.mock.json';
 import { Product } from '../models/Product';
