+++ product-service/handlers/db/productsList.mock.json
@@ -14,7 +14,7 @@
     "year": 2012,
     "engine": "2.0 Gas hybrid",
     "price": 18000,
-    "imageUrl": "https://lh3.googleusercontent.com/proxy/dDxFY7E27hZzyvhlUQFRPB2mKpT56HOSfYSWvZ_A-1B8nwQ-R844pNJic30jUV5TCxekwPu9poZhxJ3hYdxwVMVqg_V0D3yuefXB8V0",
+    "imageUrl": "https://www.motortrend.com/uploads/sites/10/2015/11/2012-toyota-rav4-sport-suv-angular-front.png",
     "title": "Toyota RAV4"
   },
   {
